requirements.txt
asgiref==3.7.2
Django==4.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
whitenoise==6.6.0
